+++ packages/core/PluginManager.ts
@@ -352,6 +352,12 @@ export default class PluginManager {
     return this.getElementTypesInGroup('track') as TrackType[]
   }
 
+  getAddTrackWorkflowElements() {
+    return this.getElementTypesInGroup(
+      'add track workflow',
+    ) as AddTrackWorkflowType[]
+  }
+
   getRpcElements() {
     return this.getElementTypesInGroup('rpc method') as RpcMethodType[]
   }
