+++ plugins/data-management/src/AddTrackWidget/components/AddTrackWidget.tsx
@@ -1,8 +1,7 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { observer } from 'mobx-react'
 import { FormControl, FormHelperText, Select, MenuItem } from '@mui/material'
-import { AddTrackWorkflowType } from '@jbrowse/core/pluggableElementTypes'
-import { getEnv, useLocalStorage } from '@jbrowse/core/util'
+import { getEnv } from '@jbrowse/core/util'
 
 // locals
 import { AddTrackModel } from '../model'
@@ -10,19 +9,17 @@ import DefaultAddTrackWorkflow from './DefaultAddTrackWorkflow'
 import PasteConfigWorkflow from './PasteConfigWorkflow'
 
 function AddTrackSelector({ model }: { model: AddTrackModel }) {
-  const [val, setVal] = useLocalStorage('trackSelector-choice', 'Default')
+  const [val, setVal] = useState('Default add track workflow')
   const { pluginManager } = getEnv(model)
-  const widgets = pluginManager.getElementTypesInGroup(
-    'add track workflow',
-  ) as AddTrackWorkflowType[]
+  const widgets = pluginManager.getAddTrackWorkflowElements()
   const ComponentMap = {
-    Default: DefaultAddTrackWorkflow,
+    'Default add track workflow': DefaultAddTrackWorkflow,
     'Add track JSON': PasteConfigWorkflow,
     ...Object.fromEntries(widgets.map(w => [w.name, w.ReactComponent])),
   } as { [key: string]: React.FC<{ model: AddTrackModel }> }
 
   // make sure the selected value is in the list
-  const val2 = ComponentMap[val] ? val : 'Default'
+  const val2 = ComponentMap[val] ? val : 'Default add track workflow'
   const Component = ComponentMap[val2]
   return (
     <>
