+++ plugins/data-management/src/AddTrackWidget/components/TextIndexingConfig.tsx
@@ -1,0 +1,134 @@
+import React, { useEffect, useState } from 'react'
+import {
+  Card,
+  CardContent,
+  IconButton,
+  InputLabel,
+  InputAdornment,
+  List,
+  ListItem,
+  Paper,
+  TextField,
+} from '@mui/material'
+import { makeStyles } from 'tss-react/mui'
+import { observer } from 'mobx-react'
+
+// icons
+import DeleteIcon from '@mui/icons-material/Delete'
+import AddIcon from '@mui/icons-material/Add'
+
+// locals
+import { AddTrackModel } from '../model'
+
+const useStyles = makeStyles()(theme => ({
+  paper: {
+    display: 'flex',
+    flexDirection: 'column',
+    padding: theme.spacing(1),
+  },
+
+  card: {
+    marginTop: theme.spacing(1),
+  },
+}))
+
+const TextIndexingConfig = observer(({ model }: { model: AddTrackModel }) => {
+  const { classes } = useStyles()
+  const [value1, setValue1] = useState('')
+  const [value2, setValue2] = useState('')
+  const [attributes, setAttributes] = useState(['Name', 'ID'])
+  const [exclude, setExclude] = useState(['CDS', 'exon'])
+  const sections = [
+    {
+      label: 'Indexing attributes',
+      values: attributes,
+    },
+    {
+      label: 'Feature types to exclude',
+      values: exclude,
+    },
+  ]
+  useEffect(() => {
+    model.setTextIndexingConf({ attributes, exclude })
+  }, [model, attributes, exclude])
+
+  return (
+    <Paper className={classes.paper}>
+      <InputLabel>Indexing configuration</InputLabel>
+      {sections.map((section, index) => (
+        <Card raised key={section.label} className={classes.card}>
+          <CardContent>
+            <InputLabel>{section.label}</InputLabel>
+            <List disablePadding>
+              {section.values.map((val: string, idx: number) => (
+                <ListItem key={idx} disableGutters>
+                  <TextField
+                    value={val}
+                    InputProps={{
+                      endAdornment: (
+                        <InputAdornment position="end">
+                          <IconButton
+                            color="secondary"
+                            onClick={() => {
+                              const newAttr = section.values.filter((a, i) => {
+                                return i !== idx
+                              })
+                              index === 0
+                                ? setAttributes(newAttr)
+                                : setExclude(newAttr)
+                            }}
+                          >
+                            <DeleteIcon />
+                          </IconButton>
+                        </InputAdornment>
+                      ),
+                    }}
+                  />
+                </ListItem>
+              ))}
+              <ListItem disableGutters>
+                <TextField
+                  value={index === 0 ? value1 : value2}
+                  placeholder="add new"
+                  onChange={event => {
+                    index === 0
+                      ? setValue1(event.target.value)
+                      : setValue2(event.target.value)
+                  }}
+                  InputProps={{
+                    endAdornment: (
+                      <InputAdornment position="end">
+                        <IconButton
+                          onClick={() => {
+                            if (index === 0) {
+                              const newAttr: string[] = attributes
+                              newAttr.push(value1)
+                              setAttributes(newAttr)
+                              setValue1('')
+                            } else {
+                              const newFeat: string[] = exclude
+                              newFeat.push(value2)
+                              setExclude(newFeat)
+                              setValue2('')
+                            }
+                          }}
+                          disabled={index === 0 ? value1 === '' : value2 === ''}
+                          color="secondary"
+                          data-testid={`stringArrayAdd-Feat`}
+                        >
+                          <AddIcon />
+                        </IconButton>
+                      </InputAdornment>
+                    ),
+                  }}
+                />
+              </ListItem>
+            </List>
+          </CardContent>
+        </Card>
+      ))}
+    </Paper>
+  )
+})
+
+export default TextIndexingConfig
